+++ frontend/app/src/components/dashboard/DashboardAssetTable.vue
@@ -270,7 +270,7 @@ watch(search, () => setPage(1));
         <AmountDisplay
           v-else
           :loading="!row.usdPrice || row.usdPrice.lt(0)"
-          no-scramble
+          is-asset-price
           show-currency="symbol"
           :price-asset="row.asset"
           :price-of-asset="row.usdPrice"