+++ frontend/app/src/types/settings/price-oracle.ts
@@ -5,7 +5,6 @@ export enum PriceOracle {
   COINGECKO = 'coingecko',
   CRYPTOCOMPARE = 'cryptocompare',
   FIAT = 'fiat',
-  MANUAL = 'manual',
   MANUALCURRENT = 'manualcurrent',
   UNISWAP2 = 'uniswapv2',
   UNISWAP3 = 'uniswapv3',